requirements.txt
Django==1.5.1
PIL==1.1.7
South==0.7.6
argparse==1.2.1
boto==2.9.0
distribute==0.6.34
dj-database-url==0.2.1
django-storages==1.1.8
easy-thumbnails==1.2
gunicorn==0.17.4
psycopg2==2.5
wsgiref==0.1.2
